added tests + base support for interp
